pyproject.toml
[tool.poetry]
name = "twiner"
version = "0.1.0"
description = "CLI that helps you with notifications for your favorite streamers on Twitch.tv."
authors = ["febits <febits@proton.me>"]
readme = "README.md"

[tool.poetry.scripts]
twiner = "twiner.cli:app"


[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
requests = "^2.31.0"
pyyaml = "^6.0.1"
plyer = "^2.1.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
isort = "^5.13.2"
pylint = "^3.0.3"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
